Move interface to JAva to avoid generics wildcard issues
